requirements.txt
autopep8==1.5.7
bcrypt==3.2.0
black==21.9b0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
gunicorn==20.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
MarkupSafe==2.0.1
mypy-extensions==0.4.3
nltk==3.6.2
numpy==1.21.2
pandas==1.3.2
pathspec==0.9.0
platformdirs==2.3.0
psycopg2-binary==2.9.1
pycodestyle==2.7.0
pycparser==2.20
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
regex==2021.8.28
requests==2.26.0
scikit-learn==0.24.2
scipy==1.7.1
six==1.16.0
sklearn==0.0
SQLAlchemy==1.4.25
threadpoolctl==2.2.0
toml==0.10.2
tomli==1.2.1
torch==1.8.0+cpu
tqdm==4.62.2
typing-extensions==3.10.0.2
urllib3==1.26.6
Werkzeug==2.0.1
yapf==0.31.0
